backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aioconsole==0.1.15
aiogram==2.3
aiohttp-socks==0.2.2
aiohttp==3.5.4            # via aiogram, aiohttp-socks
aioredis==1.2.0           # via channels-redis
appdirs==1.4.3            # via black
apscheduler==3.6.1
asgiref==3.2.2            # via channels, channels-redis, daphne
asn1crypto==0.24.0        # via cryptography
async-timeout==3.0.1      # via aiohttp, aioredis
atomicwrites==1.3.0       # via pytest
attrs==19.1.0
autobahn==19.9.2          # via daphne
automat==0.7.0            # via twisted
babel==2.7.0              # via aiogram
backcall==0.1.0           # via ipython
beautifulsoup4==4.6.0     # via wagtail
black==18.6b4
bleach==3.1.0             # via nbconvert
blinker==1.4              # via sentry-sdk
boto3==1.9.220
botocore==1.12.220        # via boto3, s3transfer
cachetools==3.1.1         # via google-auth
captcha-solver==0.1.5
certifi==2019.6.16        # via aiogram, requests, sentry-sdk
cffi==1.12.3              # via cryptography
channels-redis==2.4.0
channels==2.2.0
chardet==3.0.4            # via aiohttp, requests
click==7.0                # via black, flask
codecov==2.0.15
constantly==15.1.0        # via twisted
coverage==4.5.4           # via codecov, pytest-cov
cryptography==2.7         # via autobahn, secretstorage
cycler==0.10.0            # via matplotlib
daphne==2.3.0             # via channels
decorator==4.4.0          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
django-annoying==0.10.4
django-cors-headers==3.1.0
django-debug-toolbar==2.0
django-modelcluster==5.0  # via wagtail
django-prometheus==1.1.0
django-rest-framework==0.1.0
django-reversion-compare==0.8.6
django-reversion==3.0.4
django-taggit==1.1.0      # via wagtail
django-treebeard==4.3     # via wagtail
django==2.2.5
djangorestframework==3.10.2  # via django-rest-framework, wagtail
docutils==0.15.2          # via botocore
draftjs-exporter==2.1.6   # via wagtail
entrypoints==0.3          # via flake8, nbconvert
factory-boy==2.12.0
faker==2.0.3              # via factory-boy
feedgen==0.7.0
fire==0.2.1
flake8==3.7.8
flask==1.1.1              # via sentry-sdk, slackeventsapi, slappy
freezegun==0.3.12
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth==1.6.3
gspread==3.1.0
gunicorn==19.9.0
hiredis==1.0.0            # via aioredis
html5lib==1.0.1           # via wagtail
httplib2==0.13.1
hyperlink==19.0.0         # via twisted
idna==2.8                 # via hyperlink, requests, yarl
importlib-metadata==0.20  # via pluggy, pytest
incremental==17.5.0       # via twisted
ipykernel==5.1.2          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.8.0            # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.5.1         # via jupyter
itsdangerous==1.1.0       # via flask
jedi==0.15.1              # via ipython
jeepney==0.4.1            # via secretstorage
jinja2==2.10.1            # via flask, nbconvert, notebook
jmespath==0.9.4           # via boto3, botocore
jsonschema==3.0.2         # via nbformat
jupyter-client==5.3.1     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==6.0.0    # via jupyter
jupyter-core==4.5.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
keyring==10.3.2           # via pycookiecheat
kiwisolver==1.1.0         # via matplotlib
lxml==4.4.1               # via feedgen, python-docx
markdown==3.1.1
markupsafe==1.1.1         # via jinja2
matplotlib==3.1.1
mccabe==0.6.1             # via flake8
mistune==0.8.4            # via nbconvert
more-itertools==7.2.0     # via pytest, zipp
msgpack==0.6.1            # via channels-redis
multidict==4.5.2          # via aiohttp, yarl
mwclient==0.10.0
mypy-extensions==0.4.1    # via mypy
mypy==0.720
mysqlclient==1.4.2
nbconvert==5.6.0          # via jupyter, notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==6.0.1           # via jupyter, widgetsnbextension
numpy==1.17.1             # via matplotlib, pandas, scipy, seaborn
oauthlib==3.1.0           # via requests-oauthlib
packaging==19.1           # via pytest
pandas==0.25.1
pandocfilters==1.4.2      # via nbconvert
parso==0.5.1              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==5.4.1             # via wagtail
pluggy==0.12.0            # via pytest
prometheus-client==0.7.1  # via django-prometheus, notebook
prompt-toolkit==2.0.9     # via ipython, jupyter-console
ptyprocess==0.6.0         # via pexpect, terminado
py==1.8.0                 # via pytest
pyaes==1.6.1              # via telethon
pyasn1-modules==0.2.6     # via google-auth
pyasn1==0.4.7             # via pyasn1-modules, rsa
pycodestyle==2.5.0        # via flake8
pycookiecheat==0.4.3
pycparser==2.19           # via cffi
pycrypto==2.6.1           # via pycookiecheat
pyee==5.0.0
pyflakes==2.1.1           # via flake8
pygments==2.4.2           # via ipython, jupyter-console, nbconvert, qtconsole
pyhamcrest==1.9.0         # via twisted
pyinotify==0.9.6
pyjwt==1.7.1
pyparsing==2.4.2          # via matplotlib, packaging
git+https://github.com/berekuk/pyppeteer.git@dev
pyrsistent==0.15.4        # via jsonschema
pysocks==1.7.0
pytest-asyncio==0.10.0
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-env==0.6.2
pytest==5.1.2
python-dateutil==2.8.0
python-docx==0.8.10
python-gitlab==1.11.0
pytz==2019.2
pywatchman==1.4.1
pyzmq==18.1.0             # via jupyter-client, notebook
qtconsole==4.5.5          # via jupyter
redis==3.3.8
requests-oauthlib==1.2.0  # via mwclient
requests-toolbelt==0.9.1
requests==2.22.0
rsa==4.0                  # via google-auth, telethon
s3transfer==0.2.1         # via boto3
scipy==1.3.1              # via seaborn
seaborn==0.9.0
secretstorage==3.1.1      # via keyring
send2trash==1.5.0         # via notebook
sentry-sdk[flask]==0.11.2
six==1.12.0               # via apscheduler, autobahn, automat, bleach, captcha-solver, cryptography, cycler, faker, fire, freezegun, google-api-python-client, google-auth, html5lib, jsonschema, mwclient, packaging, prompt-toolkit, pyhamcrest, pyrsistent, python-dateutil, python-gitlab, slackclient, slackeventsapi, traitlets, txaio, wagtail, websocket-client
slackclient==1.3.2
slackeventsapi==2.1.0
slappy==0.5.1
sqlparse==0.3.0           # via django, django-debug-toolbar
telethon==1.9.0
termcolor==1.1.0
terminado==0.8.2          # via notebook
testpath==0.4.2           # via nbconvert
text-unidecode==1.3       # via faker
toml==0.10.0              # via black
tornado==6.0.3            # via ipykernel, jupyter-client, notebook, terminado
tqdm==4.35.0
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
twisted==19.7.0           # via daphne
txaio==18.8.1             # via autobahn
typed-ast==1.4.0          # via mypy
typing-extensions==3.7.4  # via mypy
tzlocal==2.0.0            # via apscheduler
unidecode==1.1.1          # via wagtail
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.25.3           # via botocore, requests, sentry-sdk
wagtail==2.7
wagtailgeowidget==5.0.2
git+https://github.com/berekuk/wagtailmath.git@master
wcwidth==0.1.7            # via prompt-toolkit, pytest
webencodings==0.5.1       # via bleach, html5lib
websocket-client==0.54.0  # via slackclient
websockets==8.0.2
werkzeug==0.15.5          # via flask
whitenoise==4.1.3
widgetsnbextension==3.5.1  # via ipywidgets
willow==1.3               # via wagtail
yarl==1.3.0               # via aiohttp
zipp==0.6.0               # via importlib-metadata
zope.interface==4.6.0     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
